+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/JsonSchema.scala
@@ -39,4 +39,6 @@ trait JsonSchema[T] {
   ): JsonSchema[_] = {
     copy(properties.transform(transformer))
   }
+
+  def findByPointer(pointer: String): Option[JsonSchema[_]] = None
 }